added 'aggs' and 'visuals'
Took the methods out of reptweet.py and put them in aggs.py instead. 
Also created a template for visuals.py as a space to put the visualization functions.
Not sure if I can/should delete reptweet.py so for now I'll just clear out the functions and leave only the imports and main-method thing.